# Conflicts:
#	server/catalog/index.cpp
#	server/catalog/inverted_index.h
#	server/catalog/table_options.h
#	server/catalog/tokenizer.cpp
#	server/catalog/tokenizer.h
#	server/connector/duckdb_physical_create_index.cpp
#	server/connector/search_filter_printer.cpp
#	server/pg/commands/create_tsdictionary.cpp
#	tests/server/connector/duckdb_search_filter_builder_test.cpp
#	tests/server/connector/duckdb_search_sink_writer_test.cpp

+++ server/catalog/table_options.h
@@ -78,8 +78,9 @@ class Column final : public Object {
     return absl::StrCat(kOffsetsNamePrefix, column_id.id());
   }
 
+  // LIST(INTEGER) -- flat offsets column: interleaved start,end pairs.
   static duckdb::LogicalType MakeOffsetsType() {
-    return duckdb::LogicalType::LIST(duckdb::LogicalType::BIGINT);
+    return duckdb::LogicalType::LIST(duckdb::LogicalType::INTEGER);
   }
 
   Column() : Object{{}, {}, {}, ObjectType::Column} {}
